api/requirements.txt
Django==2.1.3
pytz==2018.7
djangorestframework==3.9
Pillow
psycopg2-binary
# psycopg2